My submition for the drill creating a GUI
The code I used to create this GUI is based on code from an example video where we created one similar. I changed the buttons and fields and everything to adhere to the rules of the assignment. I could not for the life of me figure out how to change the font size in the title, also, I wasn't sure which font was being used in the  example so I just went with Arial.